+++ apps/admin/src/components/shared/auth/RefuseAuth.tsx
@@ -1,12 +1,17 @@
-import { authState } from '@store/auth';
 import { Navigate, Outlet } from 'react-router-dom';
-import { useRecoilValue } from 'recoil';
+import useRefresh from '@lib/hooks/auth/useRefresh';
+import { useEffect } from 'react';
 
 const RefuseAuth = () => {
-  const auth = useRecoilValue(authState);
+  const refreshToken = localStorage.getItem('refreshToken');
+  const { refreshMutation, state, setState } = useRefresh();
+
+  useEffect(() => {
+    refreshToken ? refreshMutation.mutate(refreshToken) : setState('failed');
+  }, []);
+
   // 엑세스 있으면 홈으로
-  console.log('refuseAuth', auth);
-  if (auth.accessToken) {
+  if (state === 'succeed') {
     return <Navigate replace to="/" />;
   }
   // 둘 다 없으면 로그인
